python-consumer/requirements.txt
kafka-python-ng==2.2.2
psycopg[binary]>=3.2.10
python-dotenv==1.0.0
structlog==24.1.0

# Delta Lake with PySpark
pyspark==3.5.0
delta-spark==3.2.0
pyarrow>=18.1.0
pandas>=2.2.3

# Pure Python Delta Lake (for lightweight operations)
deltalake==0.22.3

# Jupyter Notebook (for interactive queries)
jupyterlab>=4.0.0
ipykernel>=6.29.0
